pyproject.toml
[tool.poetry]
name = "reels"
version = "0.1.0"
description = ""
authors = ["Ivan Kovtun <ivan.kovtun@capgemini.com>"]
readme = "README.md"
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = ">= 3.9, < 3.12"
pytube = "^15.0.0"
certifi = "^2024.2.2"
unidecode = "^1.3.8"
faster-whisper = "^1.0.1"
ffmpeg-python = "^0.2.0"
googletrans = "^3.0.0"
moviepy = "^1.0.3"
coverage = "^7.4.4"
sqlmodel = "^0.0.18"
alembic = "^1.13.1"
gtts = "^2.5.1"
pika = "^1.3.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
ruff = "^0.4.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = [
    "F",  # Pyflakes: general errors
    "E",  # PEP8: style errors
    "W",  # PEP8: warnings
    "C90",  # Miscellaneous: custom checks
]

ignore = [
    "E501",  # Line too long
]